requirements.txt
label-studio==1.9.1.post0
python-dotenv==1.0.0
seaborn
wordcloud
beautifulsoup4
